requirements.txt
google-api-python-client==2.188.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0
gspread==6.2.1
google-auth==2.47.0
langgraph==1.0.6
google-genai==1.16.1
streamlit==1.52.2
python-dotenv==1.1.1
